Add new variables with comments
These new variables will be to add more customizing abilities for the
user. The ability to modify the size of the Shape2D of the
CollisionShape2D using absolute numbers instead of scaling the parent
Node. The other one is just a private variable to use the Shape2D
globally since it's more convenient for the upcomming changes.