fix: adjust size and direction at runtime
Since the Platform script is a tool script, it runs in the Godot Editor
and the Gameplay, but these are two different contexts and the way
Platform works depends on that. The conditional I had where I was
checking if the Node is running in the Editor or the Game was causing
the Platform to only adjust its rotation and size in the Editor and not
in the Game.

So the solution for this was to always adjust the rotation and size at
runtime no matter the context.